Added PFILE Configuration for 'Room Color' this is the ANSI color to be used when displaying the room short description.
Added color option of "NONE" which is most useful when doing this:
color BACKGROUND NONE.
This is primarily useful for those of us using transparent terminals
as what it's really doing is setting the ansi string to ANSI_NORMAL.
This way we don't get black blotches (or whatever background color)
everywhere.

Still left to do is: -----------------------------------------------------
Modfiy the color command as follows:
  color THING COLOR [BRIGHT] [BACKGROUND]

This way we'll be able to set highlighted colors (pretty) and background
color. I sort of like being able to have room names white on blue. Note:
not grey on blue (ick ick) but highlighted grey which is white.

Also:
  mobs and pc's different colors in critter list (optional of course)

And Lastly:
  some sort of % or & prepended color codes to be used in builder
  descriptions and say, emote, yell, auction, basically everywhere.
  These will be stripped (w/o affecting line wrapping) for non ANSI
  and HEGEMON clients and will be expanded (w/o affecting line wrapping)
  to ANSI escape sequences for ANSI-HAPPY clients. In addition I think
  that any input including ANSI escape sequences should be wiped clean
  when reading a clients input buf. This will ensure that nobody w/o
  ANSI support ever gets run over by those people who are currently
  kludging color support (like me) ;)

  I suppose later it would be possibly to add <COLOR></COLOR> support
  to Hegemon, but hey... that's not my baby ;)

-- Khaav