learning things like wrestling at 1% will no longer permanently affect pc's stats. Instead new methods Critter::getDEX/getDAM/etc. have been added. These methods take an optional parameter (TRUE/FALSE) to include skill bonuses to the returned value. The default is NOT to include spell bonuses. This will allow higher %lrnd to have greater benefits. This will also allow us to tweak the amount of change skills give w/o borking stats in Pfiles.
I need to finish adding these methods for all the stats, and we need to
deprecate using anything but Critter::get* methods everywhere else. i.e. stop
using things like pc.DAM which is a #define.

Along with this I updated these references for some of the stats in battle.cc

--Khaav