Added: vehicle data version tagging to db files, ticks_at_stops, in room moving and in room stopped descriptions for vehicles.
These are currently supported in OLC but are not available for modification
with one-shot commands. I plan on adding a vset command (or something to that
effect) in the near future.

I'm betting that vehicle creation with Hegemon is now officially broken.
(good thing Jaeger is working on Hegemon)

Vehicles are now visible (if they have in-room descriptions) when they are
present in the same room as a player.

I'm experimenting with the auto-exit listing. Rather than listing (??) for
non-directional exits it will list something to the effect of (portal),
(coach), etc. I think this will be a little less confusing than:
(??) N NE S SE (??) (??)

TODO:
A "destinations" or "stops" command that will tell players where a vehicle is
headed. I have yet to decide whether this will be a static text field set by
the builder or something the game figures out by itself using the vehicles
pathing information.

Start looking into player-controlled vehicles.