Made huge changes to Hegemon that I want to file for safe keeping. Most notably, just about everything in the main GUI (and a few dialog boxes/ frames) has been moved to Swing. I tried to leave logic intact and just convert to swing, but ActionEvents are going to be handled with cleaner code so that's the bulk of the actual edits.
Known problem: The complicated scrolling component of the main window
(i.e. telnet output) is heavyweight and is thus rendered on top of
menus. In the ClientDisplay.java, adding that component is commented
out so that menus are visible. So in other words, it's not a functioning
revision atm.

Jaeger