FIXES: MUD-403 Dissolving objects (corpses) placed inside any container which could present a sorted and categorized inventory listing to places were causing memory corruption and on occasion segfaults after they dissolved.
This appears to have been caused by the out_inv() routine doing bad things
that were invalidating the corpse-objects "in_list" pointers.

This was not affecting rooms (which are containers of sorts) as they did not
utilize the out_inv() function.

As a future project, objects which dissolve in player owned boxes (like the
altar and donation bin) should cause the box to be saved. This is not
currently the case.