Weaken now reduces the players strength by 10%, which depending on the player can be more or less than the previous effect. In addition, the duration of weaken has been reduced substantially.
mstat will now show the values and durations remaining in the "Affected by"
lists.

added critter::getSTR() which currently takes into account the Strength
Conditioning skill.

The contents of the file "motd" will now be sent to players after they login.

Reverted the change from the previous patch to misc.cc (<= is back to ==). It
occured to me that "-1" as a value may be of some significance and the problem
turned out to be something different.