This is for MUD-386
Though the syntax is different, room_has_critter is a command to be used in
script conditionals with the following syntax:

if room_has_critter <rnum/-1> <critter_name>

It will return true if rnum has any critters whose names match critter_name.
If rnum is passed as -1 the current room is used instead.