Fixed an order-of-operations issue in login.cc which was causing runaway HegemonHandler's.
Destructor definitions for {Hegemon,Telnet}Handler are now in their
respective .cc files.

We now properly say no when confronted with telnet linemode (rfc-1184) which
is not the same thing as kludge-linemode.

"There is an understanding known as kludge line mode which means that if
either "suppress go ahead" or "echo" is enabled but not both then telnet
operates in line at a time mode meaning that complete lines are assembled at
each end and transmitted in one "go"."

Reference:
 "The Telnet Protocol" - Peter Burden
 http://www.scit.wlv.ac.uk/~jphb/comms/telnet.html (02/26/2007)