Small functional changes to the telnet handler.
Null bytes received in text streams (not telnet options / negotiations are now
ignored. 

Added a ST_CR state; used when we receive a carriage-return in a text stream.
The rfc says we may get CR-NULL or CR-LF. CR-LF is to be treated as a single
unit (we now do so) and CR-NULL (for us) means the same thing. At least Linux
telnet uses CR-NULL's by default when SGA (suppress go ahead) is negotiated.

0x7F (delete) and 0x08 (backspace) are both treated as backspace now.

Added support for the following telnet commands:
 NOP: do nothing / no operation
 EC: erase character (untested)
 EL: erase line (untested)

Most of this is being done so that we can support non-linemode based
functionality in the future. (single-key menu options, etc)