In response to: MUD-406
Both IMMs, and Players detecting hidden, will now see hidden exits listed in
their auto-exit list.

In addition, closed exits are no longer postfixed with (closed) as that was a
bit verbose. The exit name is now simply enclosed in parens.

Hidden exits are prefixed with a * which is located inside the parens of a
closed and hidden exit.

For example: (*W) E N NW
Where west is both hidden and closed.

Note: I have not yet changed the "exits" command (the much more verbose, and
non-automatic version of auto-exits) to behave this way.