Fixed a spell message typo.
A while back the "tammuz" spell got set to skill #0 on the production port.
Turns out that a lot of ScryMUD code isn't happy with this (though some of it
doesn't care). As a result it, among other things, broke a lot of skill/spell
help. The pfile maintenance will now check and see if players know skill #0,
removing it if they do. It will then retrain any remorts with the new,
appropriate skill #.