critter::getDamGivMod() was added.
critter::getDamRecMod() and critter::getDamGivMod() now take an
include_modifiers bool. It is important to note that unlike other similar
functions, these default to true rather than false. The reasoning behind this
is that the others were in use prior to them considering things for return
value modification and getDamRecMod() was used only in very few locations,
making it easy to verify the safety of the change.

I haven't decided whether or not to invert the logic of the others, or simply
to not have a default.

The spell "ritual of power" has been added.